+++ src/lib.rs
@@ -32,6 +32,7 @@
 
 pub mod access;
 pub mod adql;
+pub mod adql_query;
 pub mod app;
 pub mod config;
 pub mod data;